refactor: move variable and argument typing rules to ANN (typing) group
