+++ tests/linter/rules/typing/missing_argument_type/test.robot
