+++ tests/linter/rules/typing/missing_argument_type/test_rule.py
