+++ tests/linter/rules/typing/missing_for_loop_variable_type/expected_output.txt
@@ -1,0 +1,5 @@
+test.robot:4:12 [I] ANN03 FOR loop variable '${index}' is missing type annotation
+test.robot:12:12 [I] ANN03 FOR loop variable '${idx}' is missing type annotation
+test.robot:22:12 [I] ANN03 FOR loop variable '${i}' is missing type annotation
+test.robot:24:16 [I] ANN03 FOR loop variable '${j}' is missing type annotation
+test.robot:35:16 [I] ANN03 FOR loop variable '${y}' is missing type annotation
