+++ tests/linter/rules/typing/missing_for_loop_variable_type/test.robot
