+++ tests/linter/rules/typing/missing_section_variable_type/test_rule.py
